fix: change getPercentual return

+++ src/main/java/dev/thiagorodrigues/livraria/application/dtos/RelatorioLivrosPorAutor.java
@@ -1,11 +1,13 @@
 package dev.thiagorodrigues.livraria.application.dtos;
 
+import java.math.BigDecimal;
+
 public interface RelatorioLivrosPorAutor {
 
     String getAutor();
 
     Long getQuantidade();
 
-    Double getPercentual();
+    BigDecimal getPercentual();
 
 }
